Add files via upload
Welcome to my Power BI Data Analyst Guided Project, a crucial step in my journey to establish a career in data analysis. In this repository, I've documented my exploration of real-world data and the development of insightful reports using Power BI.This guided project reflects my passion for data analysis and my commitment to acquiring the skills required to succeed in this dynamic field. It's a glimpse into my proficiency with Power BI and my ability to extract actionable knowledge from data.